requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
streamlit
requests
pandas
beautifulsoup4